fix(platform): share arg assertion helpers across build platforms
assertContains/assertNotContains lived in mq_wsl_coverage_test.go, which is gated behind //go:build windows, so the unconstrained tmux_test.go failed to typecheck on Linux CI with 'undefined: assertContains'. Move the shared helpers into an unconstrained assert_test.go so both the Unix tmux tests and the Windows WSL tests can use them.
